Update Python 3 compatibility.
Run the hook loader (and all Python-based hooks) within the WORKON_HOME directory, instead of wherever the shell happened to be before. In particular, this effects projects templates, which used to be run in the project directory. That path is now passed as an argument to the template, instead.

Update tests to account for differences in output when exceptions are reported between different Python versions.

Use typeset to properly scope variables used in deactivate().

Fix syntax for declaration of mktmpenv.